# swagger.yml
openapi: 3.0.0
info:
  title: Web Calendar API
  description: API для управления событиями и пользователями в веб-календаре.
  version: 1.0.0
servers:
  - url: /api
paths:
  /register:
    post:
      summary: Регистрация нового пользователя
      operationId: src.registration.registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: user123
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: password123
      responses:
        201:
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  username:
                    type: string
                    example: user123
                  email:
                    type: string
                    example: user@example.com
        400:
          description: Аккаунт с такой электронной почтой уже существует
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Аккаунт с такой электронной почтой уже существует.
  /login:
    post:
      summary: Вход пользователя
      operationId: src.registration.entrance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: password123
      responses:
        200:
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    example: 1
                  username:
                    type: string
                    example: user123
                  email:
                    type: string
                    example: user@example.com
                  password:
                    type: string
                    example: hashed_password123
        400:
          description: Неверный адрес электронной почты или пароль
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Неверный адрес электронной почты или пароль.
...